Silicon/Hisilicon: Use ReportStatusCodeLib (DXE) in MdeModulePkg
REF:https://bugzilla.tianocore.org/show_bug.cgi?id=1800

This commit will update the DSC.INC file to consume the
ReportStatusCodeLib (DXE phase instance) in MdeModulePkg, instead of using
the one in IntelFrameworkModulePkg.

Please note that, the difference between the 2 ReportStatusCodeLib are:

A. The MdeModulePkg one drops the support of
   FRAMEWORK_EFI_RUNTIME_SERVICES;
B. The MdeModulePkg one adopts a fix to avoid using AllocatePool().
   (edk2 commit c22f52c5e79b9782648576efb8382bb04da60b5b)

Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Cc: Leif Lindholm <leif.lindholm@linaro.org>
Cc: Michael D Kinney <michael.d.kinney@intel.com>
Signed-off-by: Hao A Wu <hao.a.wu@intel.com>