Platform/BeagleBoard: Use ReportStatusCodeLib in MdeModulePkg
REF:https://bugzilla.tianocore.org/show_bug.cgi?id=1800

This commit will update the DSC file to consume the ReportStatusCodeLib
(DXE phase and Runtime instances) in MdeModulePkg, instead of using the
one in IntelFrameworkModulePkg.

Please note that, the DXE and Runtime ReportStatusCodeLib instances are
almost identical between IntelFrameworkModulePkg and MdeModulePkg. Thus,
there is no functional impact for this commit.

Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Cc: Leif Lindholm <leif.lindholm@linaro.org>
Cc: Michael D Kinney <michael.d.kinney@intel.com>
Signed-off-by: Hao A Wu <hao.a.wu@intel.com>